[add] vector and matrix modules to separate implementations.
